DATA-298B/requirements.txt
chromadb==0.4.18
sentence-transformers==2.2.2
boto3>=1.34.0
python-dotenv==1.0.0
